refactor: 지역, 카테고리, 인기상품 package 경로 변경
