+++ src/main/java/com/example/secondhand/domain/interest_product/dto/ReadInterestProductListDto.java
@@ -1,4 +1,4 @@
-package com.example.secondhand.domain.product.dto;
+package com.example.secondhand.domain.interest_product.dto;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -6,7 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-public class ReadPopularProductListDto {
+public class ReadInterestProductListDto {
 
 	@Getter
 	@Setter
