+++ src/main/java/com/example/secondhand/domain/interest_product/dto/ReadPopularProductListDto.java
@@ -1,22 +1,18 @@
-package com.example.secondhand.domain.product.dto;
+package com.example.secondhand.domain.interest_product.dto;
 
-import javax.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-public class AddInterestProductDto {
+public class ReadPopularProductListDto {
 
 	@Getter
 	@Setter
 	@AllArgsConstructor
 	@NoArgsConstructor(access = AccessLevel.PROTECTED)
 	public static class Request {
-
-		@NotNull
-		private Long productId;
-
+		private int page = 0;
 	}
 }
